backend/requirements.txt
flask
flask_cors
pymongo
dnspython
flask_jwt_extended
python-dotenv
